Rename module fileio_registry to filesystem_registry
